@@ -1,13 +1,14 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
 
-import { fetchEntries, addEntry } from "../store/entries/settings";
+import { fetchEntries } from "../store/entries/settings";
 
 import SideMenu from "./SideMenu";
 import Entries from "./Entries";
 
 class App extends Component {
-  componentWillMount() {
+  
+  componentDidMount() {
     this.props.fetchEntries();
   }
 
